pyproject.toml
[project]
name = "alpaca-mcp-server"
version = "0.1.0"
description = "FastMCP server implementation for Alpaca Trading API"
authors = [
    { name = "embeddednature", email = "isaac@embeddednature.com" },
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.8.0",
    "alpaca-py>=0.30.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["alpaca_mcp_server"] 

[dependency-groups]
dev = [
    "fastmcp[dev]>=2.8.1",
]
